requirements.txt
Django>=6.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
Pillow>=10.2.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
gunicorn>=22.0.0
requests
